move report functions to report module
